refactor(TMS-31): fixing the map component location
